Turn ship to face direction of travel
React does not support SVG Image elements. We need to do this ourselves.
In order to do this, we need to use 'dangerouslySetInnerHTML' which is a
React method.

If you display an image at the coordinates of the ship, the image will
start at that point, meaning that it will be offset. You need to adjust
the image to sit right on top of the ship coordinates. Also, SVG needs
to know the exact point in which to initiate it's rotation. Figuring out
the center of the image takes a bit of trigonometry to adjust the image
properly.

Also, figuring out the angle of rotation takes a little bit of
trigonometry as well.